+++ src/js/app/main/index.js
@@ -2,22 +2,27 @@
  * @author rubyisapm
  */
 import Vue from 'vue';
-import Textarea from '../../../../index';
+import Textarea from 'main';
 new Vue({
     el: '#app',
     components: {
         'ct-adc-textarea': Textarea
     },
     data: {
-        limitedLength: 100,
-        cut: false
+        value: 'helloworldefefeffe',
+        limitedLength: 10,
+        cut: false,
+        replaceByEmpty: false
     },
     methods: {
         changePass(pass){
-            console.log(pass);
+            console.log('=============change pass================', pass);
         },
         change(content){
-            console.log(content);
+            console.log('===========change============', content);
+        },
+        input(content){
+            console.log('================input==================', content);
         }
     }
 });