+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## 更新日志
 
+### 2.0.0-alpha.1
+
+*2018-04-06*
+
+- 优化 加入v-model替代原来的初始值配置和手动同步
+
+- 修改 将limited-length改为maxlength
+
+- 新增 新增事件响应如input change blur focus
+
 ### 1.0.0
 
 *2017-12-12*